embedder/requirements.txt
face_recognition==1.3.0
faiss-gpu==1.7.2
numpy==1.26.4
pillow==10.2.0
sqlite3-api==0.1.0
qdrant-client==1.7.4
watchdog==3.0.0
retina-face==0.0.17
opencv-python-headless==4.9.0.80
deepface==0.0.79
tensorflow==2.15.0
mtcnn==0.1.1
scikit-image==0.22.0
torch==2.2.1
torchvision==0.17.1 